chore(scripts): Exclude all modules from dependencies in bundles
